Back to publishing when creating a release
While it's nice to be able to push a tag and then get the GitHub
Actions CI to build the project, publish its Maven package to GitHub,
and create the release with release notes, it is not ideal.
In particular, the created release does not contain a link to the
Maven package that was published. Instead, I'll go back to the old
process, where I create a tag, create a Release in GitHub, and this
triggers the build and publish of the Maven package. Somehow, this
also associates the package with the release correctly.